requirements.txt
tensorflow-cpu
numpy
pandas
scikit-learn
flask
streamlit
openai
requests